Add Generator Code That Guarantees Solvability
The RandomGenerator will repeatedly generate random
levels until it generates one that is solvable. A
solvable level is one in which all required collectables
and the final goal are reachable by the player.

This doesn't allow for any difficult levels that have
challenges since everything must be directly reachable,
but it does at least make the game playable!