Add Flippers and Monsters
I added monsters to the level. For PuzzleScript, you
can just map the monsters to empty tiles so they don't exist
in that game.

I also added flippers so the player can swim. I updated the
solver so that it knows whether you have flippers so that it
can tell if spots are reachable with/without flippers.

I also added a barrier in front of the goal that only opens
if you collect all the collectables.