Add Graph Unraveler
Graph Unraveler simulates the dependency graph by treating edges as
springs and nodes as repulsive particles. Movement of the graph's
edges is simulated until it comes to rest. This appears to be a
fairly effective way to unravel a graph and ensure nodes are spread
out evenly and that there are few overlaps of edges.