Add force when edges overlap, add max step timeout
Added an extra force that pushes on nodes when their edges overlap.
It isn't perfect, but I think it helps a bit. There is still an
issue where overlapping edges can be a more stable state than
non-overlapping edges. I'm not sure how to recognize that. It
occurs when the overlapping state puts less stress on the spring
edges than the non-overlapping state, such as in the unravel test
number 2 which often ends with overlaps. Perhaps we could have
dynamically lengthed edges which change if there is too much stress
over too much time? Sounds complicated...

I also added a timeout when the simulation runs for too long and
doesn't reach a stable state. This is just so we don't keep
simulating forever.