Split Unraveler from Spatial Generator
The process of unraveling a graph is only part of the work
that is needed to convert a mission graph to a spatial graph.
I separated the unraveling from te rest of the conversion
process. Specifically, creating points and adjacency matrix
from mission graph, then centering the unraveled graph in
the level and aligning the positions with tiles in the
level grid.