Make Graph Node's Add/Remove Symmetric
Previously, given two nodes A and B, adding B
as a child of A did not automatically add
A as a parent of B. This made it hard to traverse
the mission graph both directions unless you
specifically added both relationships each time,
which was tedious and error-prone. Now when you
add B as a child of A, A is automatically added
as a parent of B. The same holds for removing
relationships as well.