Generate random levels incrementally
Random levels were previously being generated
all in one shot, and if they weren't solvable, they
were thrown away and regenerated from scratch. Now,
levels are generated by adding one element at a time
and seeing if it works, if not, that element is moved
around the level until it is placed in a spot that
is solvable and non-trivial. If there is no place,
then the map is regenerated from scratch.