Change mission graph on failure, add random blocks
If we try all the different combinations and can't generate
a solvable level, don't just scrap the physical layout of the
level, but also scrap the mission graph in case it happened to
generate one that was too hard/impossible for the solver.
I think there were cases where I thought it was just taking too
long to find a solution, when actually a graph had been
generated which the solver couldn't figure out, so it looped forever.
Changing the mission should break that loop.

Also, I added some random wall tiles to add some noise to the
levels and make them look more interesting than just plain rooms.