Add Support for Different Colored Keys!
Now instead of the keys being all red, they can be
red, blue, green, and yellow. This makes for much more
interesting levels because a door might lead to an empty room
with just a key in it, then you have to head somewhere else
to open the door that belongs to that key. Before, it would
just be a simple linear path for the most part.