Add Lock as Child to Key When Adding as Lock
When you add a lock to a key (or visa-versa) by doing
key.add_lock_s() or lock.add_key_s(), also add the lock
as a child of the key and the key as a parent to the lock.
Also, do this when passing params to init().

This is a nice shortcut because we need to add those links
anyway and this saves us a step each time we add a key/lock
pair.