Change Mission Generator to be Recursive
Previously, the generator would add mission nodes one at a time
and would try several different placements for the node. If it
exhausted all of its placements, it would totally scrap the level
and start afresh. Now, it backtracks one node and continues from
there in a recursive fashion. It only gives up on a level after
it has backtracked through all the nodes to the beginning.