Improve Graph Gen Readability
It was kind of hard to understand what was happening in the graph
generation code since variables were given single letter names.
I also broke functions down into more bite-sized, reusable pieces
so that it is easier to tell what each piece of code is doing.

I made all these changes because I was trying to get the graph
generation code to work using sets instead of lists, but I couldn't
tell what was happening in order to fix it.