Remove max_depth for get_random_descendant/ancestor
I felt like setting a maximum depth of 3 for going up or down the tree
was just an arbitrary number, and it wouldn't scale well if we wanted to
build bigger graphs at some point. Since it is impossible to determine
how many steps to take before traversing the graph, I modified the
methods to find ALL descendants/ancestors and then pick one of them at
random. I feel like this was more of the intent in the first place
anyway.