Change Graph Gen to not use hardcoded values
Graph Generator had hardcoded [1, 2, 3, 4] which is shorter to type than
alternatives, but makes the code brittle and hard to modify. I wanted
to change the branch_probabilities aesthetic to something other than 4
branch sizes and had to change that line. So instead, I changed the code
to build a range based on the size of the branch_probabilities array.