Change Sokoban Goal from Water to Bomb
There was an issue where if Flippers were in a level with Sokoban
puzzles, the Sokoban puzzles could be completely skipped because the
player could just swim through the water with the flippers. We instead
use bombs as the Sokoban goals so the block will blow up the bomb. And
flippers can still be used for their own thing.